feat(vue-demo): create translations helpers
