+++ packages/cms-base/components/SwProductAddToCart.vue
@@ -5,16 +5,15 @@ import {
   useNotifications,
   useCartNotification,
   useCartErrorParamsResolver,
-  useCmsTranslate,
 } from "#imports";
 import { toRefs } from "vue";
 import { defu } from "defu";
 import type { Schemas } from "#shopware";
+import { getCmsTranslate } from "@shopware-pwa/helpers-next";
 
 const { pushSuccess, pushError } = useNotifications();
 const { getErrorsCodes } = useCartNotification();
-const { cmsT } = useCmsTranslate();
-
+const { resolveCartError } = useCartErrorParamsResolver();
 const props = defineProps<{
   product: Schemas["Product"];
 }>();
@@ -51,8 +50,8 @@ const addToCartProxy = async () => {
   await addToCart();
   const errors = getErrorsCodes();
   errors?.forEach((element) => {
-    const { messageKey, params } = useCartErrorParamsResolver(element);
-    pushError(cmsT(translations.errors[messageKey], params));
+    const { messageKey, params } = resolveCartError(element);
+    pushError(getCmsTranslate(translations.errors[messageKey], params));
   });
 
   if (!errors.length)
