+++ packages/composables/src/useCart.ts
@@ -84,7 +84,6 @@ export type UseCartReturn = {
    * Get cart errors
    */
   consumeCartErrors(): Schemas["Cart"]["errors"];
-  getItem(id: string): Schemas["LineItem"] | undefined;
 };
 
 /**
@@ -279,10 +278,6 @@ export function useCartFunction(): UseCartReturn {
     return errors;
   };
 
-  const getItem = (id: string) => {
-    return cartItems.value.find((item) => item.id === id);
-  };
-
   return {
     addProduct,
     addProducts,
@@ -300,7 +295,6 @@ export function useCartFunction(): UseCartReturn {
     isEmpty,
     isVirtualCart,
     consumeCartErrors,
-    getItem,
   };
 }
