+++ packages/composables/src/useCartErrorParamsResolver.ts
@@ -1,8 +1,10 @@
 import { useCart } from "#imports";
 
 export type UseCartErrorParamsResolver = {
-  params: { [key: string]: string | number | null } | null;
-  messageKey: string;
+  resolveCartError: (errorObject: ErrorObjectParam) => {
+    params: { [key: string]: string | number | null } | null;
+    messageKey: string;
+  };
 };
 
 export type ErrorObjectParam = {
@@ -13,45 +15,52 @@ export type ErrorObjectParam = {
   level?: number;
 };
 
-export function useCartErrorParamsResolver(
-  errorObject: ErrorObjectParam,
-): UseCartErrorParamsResolver {
-  let params = null;
-  let messageKey = errorObject.messageKey;
+export function useCartErrorParamsResolver(): UseCartErrorParamsResolver {
+  const resolveCartError = (errorObject: ErrorObjectParam) => {
+    let params = null;
+    let messageKey = errorObject.messageKey;
 
-  const { getItem } = useCart();
+    const { cartItems } = useCart();
 
-  const extractProductId = (message: string, error: string) => {
-    return message.replace(error, "");
-  };
+    const extractProductId = (message: string, error: string) => {
+      return message.replace(error, "");
+    };
 
-  const buildProductStockReached = () => {
-    const productId = extractProductId(
-      errorObject.key,
-      "product-stock-reached",
-    );
-    const product = getItem(productId);
+    const buildProductStockReached = () => {
+      const productId = extractProductId(
+        errorObject.key,
+        "product-stock-reached",
+      );
+      const product = getItem(productId);
+
+      return {
+        name: product?.label || "",
+        quantity: product?.quantityInformation?.maxPurchase || null,
+      };
+    };
+
+    const getItem = (id: string) => {
+      return cartItems.value.find((item) => item.id === id);
+    };
+
+    switch (errorObject.messageKey) {
+      case "product-stock-reached":
+        params = buildProductStockReached();
+        if (!params.name || !params.quantity) {
+          messageKey = "product-stock-reached-empty";
+          params = null;
+        }
+        break;
+      default:
+        params = { ...errorObject };
+    }
 
     return {
-      name: product?.label || "",
-      quantity: product?.quantityInformation?.maxPurchase || null,
+      params,
+      messageKey,
     };
   };
-
-  switch (errorObject.messageKey) {
-    case "product-stock-reached":
-      params = buildProductStockReached();
-      if (!params.name || !params.quantity) {
-        messageKey = "product-stock-reached-empty";
-        params = null;
-      }
-      break;
-    default:
-      params = { ...errorObject };
-  }
-
   return {
-    params,
-    messageKey,
+    resolveCartError,
   };
 }
